feat: transitions-gui prototype
