+++ pyproject.toml
@@ -116,6 +116,8 @@ black = "^23.3.0"
 pytest = "^7.3.2"
 pytest-cov = "^4.1.0"
 pytest-asyncio = "^0.21.0"
+transitions-gui = "^0.1.0"
+tornado = "^6.3.2"
 
 [tool.poetry.scripts]
 myoktros = "myoktros:entrypoint"
